feat (6) selectform 분리, lectures미리 패칭
